+++ app/src/main/java/com/example/testt/data/ReminderDatabase.kt
@@ -1,7 +1,8 @@
-package com.example.testt.data.model
+package com.example.testt.data
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import com.example.testt.data.model.ReminderDBO
 
 @Database(
     entities = [ReminderDBO::class],
